pyproject.toml
[tool.poetry]
name = "GameEngine"
version = "0.1.0"
description = ""
authors = ["rrodrig6 <rrodrig6@uccs.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
pygame = "^2.1.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
